+++ src/main/java/net/evenh/versionmonitor/application/hosts/github/GitHubProject.java
@@ -1,17 +1,15 @@
 package net.evenh.versionmonitor.application.hosts.github;
 
-import net.evenh.versionmonitor.domain.projects.AbstractProject;
-import net.evenh.versionmonitor.domain.projects.Project;
-
 import javax.persistence.Entity;
 import javax.persistence.Table;
+import net.evenh.versionmonitor.domain.projects.Project;
 
 /**
  * Represents a GitHub software project.
  */
 @Entity
 @Table(name = "projects_github")
-public class GitHubProject extends AbstractProject implements Project {
+public class GitHubProject extends Project {
   public GitHubProject() {
   }
