+++ src/main/java/net/evenh/versionmonitor/domain/projects/ProjectRepository.java
@@ -16,16 +16,16 @@ import java.util.Optional;
  * @since 2016-01-10
  */
 @Repository
-public interface ProjectRepository extends JpaRepository<AbstractProject, Long> {
-  Optional<AbstractProject> findByIdentifier(String identifier);
+public interface ProjectRepository extends JpaRepository<Project, Long> {
+  Optional<Project> findByIdentifier(String identifier);
 
   /**
    * Finds a project by supplying a release.
    *
    * @param release A <code>Release</code> object.
-   * @return An Optional<code>AbstractProject</code> for describing whether a project was found or
+   * @return An Optional<code>Project</code> for describing whether a project was found or
    *         not.
    */
-  @Query("select a from AbstractProject a inner join a.releases r where r = :r")
-  Optional<AbstractProject> findByRelease(@Param(value = "r") Release release);
+  @Query("select a from Project a inner join a.releases r where r = :r")
+  Optional<Project> findByRelease(@Param(value = "r") Release release);
 }
