+++ src/main/java/net/evenh/versionmonitor/domain/projects/ProjectService.java
@@ -14,14 +14,14 @@ public class ProjectService {
   /**
    * Finds all projects persisted in the database.
    */
-  public List<AbstractProject> findAll() {
+  public List<Project> findAll() {
     return repository.findAll();
   }
 
   /**
    * Finds a project by id.
    */
-  public Optional<AbstractProject> findOne(Long id) {
+  public Optional<Project> findOne(Long id) {
     return Optional.ofNullable(repository.findOne(id));
   }
 
@@ -35,21 +35,21 @@ public class ProjectService {
   /**
    * Persists a project to database.
    */
-  public AbstractProject persist(AbstractProject project) {
+  public Project persist(Project project) {
     return repository.saveAndFlush(project);
   }
 
   /**
    * Deletes a project.
    */
-  public void delete(AbstractProject project) {
+  public void delete(Project project) {
     repository.delete(project);
   }
 
   /**
    * Finds a project by a given release.
    */
-  public Optional<AbstractProject> findByRelease(Release release) {
+  public Optional<Project> findByRelease(Release release) {
     return repository.findByRelease(release);
   }
 }
